+++ app/src/main/java/com/toyprojects/card_pilot/model/Transaction.kt
@@ -5,7 +5,7 @@ data class Transaction(
     val merchant: String,
     val date: String,     // "MM.dd"
     val time: String,     // "HH:mm"
-    val amount: Double,
-    val eligible: Double? = null,  // Benefit 모델로 옮겨야 함
+    val amount: Long,
+    val eligible: Long? = null,  // Benefit 모델로 옮겨야 함
     val monthGroup: String // "2024년 2월"
 )
